fix: add file name to logs (log file source) (#61)
Co-authored-by: Charity Kathure <ckathure@microsoft.com>

fixes #23 

The log file messages only contain the contents of the log files with no indication of the source of the content. This makes it hard to know where these messages are coming from. (Issue: #23 )

Some changes are required to include the file names when displaying contents of log files.
In the case of event logs, the message source is indicated in the Channel tag (application / system):