hack: dockerfiles: start off a nanoserver plus image
the official mcr.microsoft.com/windows/nanoserver:ltsc2022 image
does not ship with some binaries so as to reduce the image size.
Some of those are needed for testing, etc.

We are therefore creating a custom nanoserver aka nanoserver-plus
image that ships with these binaries, and more others as needed

Signed-off-by: Anthony Nandaa <profnandaa@gmail.com>